rewrite of culture action link taghelper
* Added better parameter checking support when parsing
LocalizedRouteConvention which it adds to the CultureActionData if
possible for the LocalizationDataHandler at a later time.
* Using that in CultureActionLinkTagHelper to get the correct route +
parameters.
-- It requires more testing.

* Added GetOrderedParameters() that returns the RouteValues sorted
* Also updated the code in LocalizationDataHandler because apperently it
was using the a_ prefix not just for the arguments!

* Added LocalizedHTMLGenerator but I'm not sure if it should be used
yet, but it serves as research data into the asp.net core codebase for
now.
-- I was thinking it can be used if you don't want to write cms-culture
at all and have it by default overwrite all links.

Tests:
Updated _Layout.cshtml so it uses both asp-route- and asp-all-route-data
approach.

Updated tests for LocalizedRouteConvention and added a class so write
the input & expected result on same line!